Move things around 10 / 12
Split Evaluator into Evaluator / EvaluatorImpl
