+++ amm/src/main/scala/ammonite/util/package.scala
@@ -1,0 +1,5 @@
+package ammonite
+
+package object util {
+  type ClassFiles = Vector[(String, Array[Byte])]
+}
